+++ frontend/src/views/ProductView.vue
@@ -38,46 +38,7 @@
       <div>{{ user.gender }}</div>
     </div>
     <div class="product-detail">{{ product.description }}</div>
-    <CommonModalVue
-      :is-visible="isVisible"
-      @updateVisible="(status: boolean) => (isVisible = status)"
-    >
-      <div class="modal">
-        <div class="modal-title">{{ ORDER_MODAL.TIME }}</div>
-        <div class="modal-date">
-          <div class="modal-text">
-            <p v-for="(text, index) in ORDER_MODAL.CONTENTS" :key="index">
-              {{ text }}
-            </p>
-          </div>
-          <div class="modal-input">
-            <span>{{ ORDER_MODAL.DATE }} </span>
-            <input
-              type="date"
-              :value="dateFormat(wantedDate[0])"
-              @change="onChangeDate($event, 0)"
-            />
-            ~
-            <input
-              type="date"
-              :value="dateFormat(wantedDate[1])"
-              @change="onChangeDate($event, 1)"
-            />
-          </div>
-          <div class="modal-input">
-            <span>{{ ORDER_MODAL.TIME }} </span>
-            <input type="time" />
-            ~
-            <input type="time" />
-          </div>
-        </div>
-        <div class="modal-requirement">
-          <label for="requirement"> {{ ORDER_MODAL.REQUIREMENT }}</label>
-          <textarea name="" id="" cols="30" rows="10"></textarea>
-        </div>
-        <button @click="onClickOrderSubmit">{{ ORDER_MODAL.SUBMIT }}</button>
-      </div>
-    </CommonModalVue>
+    <OrderModal v-model:is-visible="isVisible" :product-id="0" />
   </div>
 </template>
 
@@ -85,12 +46,10 @@
 import router from '@/router';
 import { ref } from 'vue';
 import { useRoute } from 'vue-router';
+import OrderModal from '@/components/OrderModal.vue';
 import image from '@/assets/tmp/images/image.png';
 import banner1 from '@/assets/tmp/images/banner1.jpeg';
-import { ORDER_MODAL, PRODUCT } from '@/constants/strings/product';
-import CommonModalVue from '@/components/CommonModal.vue';
-import { dateFormat } from '@/utils/format';
-import { compareDate } from '@/utils/date';
+import { PRODUCT } from '@/constants/strings/product';
 
 const route = useRoute();
 
@@ -114,7 +73,6 @@ const user = ref({
   gender: 'gender'
 });
 const isVisible = ref(true);
-const wantedDate = ref([new Date(), new Date()]);
 
 const onClickBannerBtn = (flag: string) => {
   if (flag === 'next') {
@@ -137,24 +95,6 @@ const onClickReport = () => {
 
 const onClickOrder = () => {
   isVisible.value = true;
-};
-
-const onChangeDate = (event: Event, index: number) => {
-  const date = (event.target as HTMLInputElement).value;
-  wantedDate.value[index] = new Date(date);
-};
-const onClickOrderSubmit = () => {
-  console.log(wantedDate.value, new Date());
-  const currentDate = new Date();
-  if (
-    compareDate(currentDate, wantedDate.value[0]) < 0 ||
-    compareDate(wantedDate.value[0], wantedDate.value[1]) < 0
-  ) {
-    alert('날짜를 정확히 입력해주세요');
-    return;
-  }
-  isVisible.value = false;
-  // TODO: 주문서 제출
 };
 </script>
 
@@ -286,78 +226,6 @@ const onClickOrderSubmit = () => {
   margin-bottom: 42px;
   padding: 24px;
 }
-.modal {
-  width: 100%;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  gap: 12px;
-
-  font-size: 24px;
-  overflow: auto;
-}
-.modal > .modal-title {
-  width: 100%;
-  border-bottom: 1px solid black;
-  font-size: 48px;
-  font-weight: 700;
-}
-.modal-date {
-  display: flex;
-  flex-direction: column;
-}
-.modal-date > .modal-text {
-  margin: 12px 0;
-}
-.modal-input {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 12px;
-  margin-top: 12px;
-}
-.modal-input > span {
-  width: 60px;
-}
-.modal > .modal-requirement {
-  width: 60%;
-  display: flex;
-  flex-direction: column;
-}
-.modal-requirement > label {
-  font-size: 16px;
-  margin-bottom: 6px;
-}
-.modal-requirement > textarea {
-  padding: 12px;
-  font-size: 18px;
-}
-.modal > button {
-  margin-top: 12px;
-  padding: 12px 18px;
-
-  border: 1px solid rgba(130, 130, 130, 0.732);
-  box-shadow: 1px 1px 10px rgba(145, 145, 145, 0.524);
-}
-.modal > button:active {
-  background-color: rgb(219, 219, 219);
-}
-
-input[type='date'],
-input[type='time'] {
-  background-color: skyblue;
-  padding: 13px;
-  font-family: 'Roboto Mono', monospace;
-  color: #ffffff;
-  font-size: 16px;
-  border: none;
-  outline: none;
-  border-radius: 5px;
-}
-
 @media screen and (max-width: 1023px) {
   .info {
     flex: 1;
